worker/requirements.txt
pymongo
sentence-transformers
flask
requests
python-docx
pypdf
python-dotenv
